Moved `status` to `app` as a sub-command.
